pyproject.toml
[tool.poetry]
name = "django-next-ninja"
version = "0.1.0"
description = ""
authors = ["Osmar Pérez Bautista <osmarpb.97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-ninja = "^0.17.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
